restructure repo; add docs
