+++ src/attention_is_all_you_need/__init__.py
@@ -1,5 +1,5 @@
 from .blocks import DecoderBlock, EncoderBlock
-from .transformer import Decoder, Encoder, Transformer
+from .transformer import EncoderDecoderTransformer
 from .multihead_attention import MultiHeadAttention
 from .position_wise_ffn import PositionwiseFeedForward
 from .positional_encoding import PositionalEncoding
