+++ src/attention_is_all_you_need/scaled_dot_product_attention.py
@@ -10,20 +10,20 @@ class ScaledDotProductAttention(nn.Module):
     r"""ScaledDotProductAttention (Section 3.2.1 of paper).
 
     Args:
-        embedding_size (`int`):
-        temperature (`float`, *optional*):
+        embedding_dim (`int`):
+            The dimension of the embedding space (d_model in paper).
     """
 
-    def __init__(self, query_key_size: int) -> None:
+    def __init__(self, query_key_dim: int) -> None:
         super().__init__()
 
         # In the original paper, product of query and key_T are normalized by square root of
         # embedding size. Here, we allow for normalizing with a temperature value too. If
-        # temperature is not `None`, it will be used. Otherwise, square root of `embedding_size`
+        # temperature is not `None`, it will be used. Otherwise, square root of `embedding_dim`
         # will be used.
-        self.query_key_size = query_key_size
+        self.query_key_dim = query_key_dim
 
-        scale = torch.sqrt(torch.FloatTensor([query_key_size]))
+        scale = torch.sqrt(torch.FloatTensor([query_key_dim]))
         self.register_buffer("scale", scale)
         self.scale: T
 
@@ -31,8 +31,8 @@ class ScaledDotProductAttention(nn.Module):
 
     def forward(self, query: T, key: T, value: T, mask: Optional[T] = None) -> T:
         # 1. MatMul
-        #  query: [batch_size, num_heads, seq_length, query_key_size]
-        #  key_T: [batch_size, num_heads, query_key_size, seq_length]
+        #  query: [batch_size, num_heads, seq_length, query_key_dim]
+        #  key_T: [batch_size, num_heads, query_key_dim, seq_length]
         # result: [batch_size, num_heads, seq_length, seq_length]
         key_T = key.transpose(2, 3)
         x = torch.matmul(query, key_T)
@@ -49,8 +49,8 @@ class ScaledDotProductAttention(nn.Module):
 
         # 5. MatMul
         #      x: [batch_size, num_heads, seq_length, seq_length]
-        #  value: [batch_size, num_heads, seq_length, value_size]
-        # result: [batch_size, num_heads, seq_length, value_size]
+        #  value: [batch_size, num_heads, seq_length, value_dim]
+        # result: [batch_size, num_heads, seq_length, value_dim]
         x = torch.matmul(x, value)
 
         return x
